--- framework/src/main/java/annotation/GET.java
@@ -1,12 +1,0 @@
-package main.java.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface GET {
-    String value() default "";
-}
