--- framework/src/main/java/annotation/JsonIgnore.java
@@ -1,11 +1,0 @@
-package main.java.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface JsonIgnore {
-}
