--- framework/src/main/java/annotation/PathParam.java
@@ -1,12 +1,0 @@
-package main.java.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface PathParam {
-    String value() default "";
-}
