+++ framework/src/main/java/com/encora/framework/annotation/App.java
@@ -1,0 +1,12 @@
+package com.encora.framework.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface App {
+    String value() default "";
+}
