+++ framework/src/main/java/com/encora/framework/annotation/JsonIgnore.java
@@ -1,0 +1,11 @@
+package com.encora.framework.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface JsonIgnore {
+}
