+++ framework/src/main/java/com/encora/framework/exception/ServerErrorException.java
@@ -1,0 +1,4 @@
+package com.encora.framework.exception;
+
+public class ServerErrorException {
+}
