--- framework/src/main/java/exception/ServerErrorException.java
@@ -1,4 +1,0 @@
-package main.java.exception;
-
-public class ServerErrorException {
-}
